Replace usage of outdated Bio::Root:Test (issue #3)
The main issue is the use of test_skip to check for network connection
which was meant to be used when this was part of the core bioperl
distribution and it was using Module::Build.  This is no longer the
cases so its use was replaced with Test::RequiresInternet after moving
the associated tests to their own test files.

Also took the opportunity to sort the issue of high code duplication
and that the refseq for get_sequence block was actually connecting to
genbank (although even after changing it to refseq the test continues
to fail).